reasoning_trace/requirements.txt
boto3==1.34.34
python-dotenv==1.0.1
networkx==3.2.1
pydantic==2.6.1 